requirements.txt
# ============================================================================
# FashionMatch - Requirements
# ============================================================================
# Python 3.10-3.12 required (ChromaDB compatibility constraint)
# After installation, run: playwright install chromium

# Core ML Framework
# ============================================================================
torch>=2.0.0,<3.0.0
torchvision>=0.15.0,<1.0.0

# AI Models
# ============================================================================
# CLIP via Hugging Face Transformers
transformers>=4.35.0,<5.0.0

# DINOv2 loaded via torch.hub - no separate package needed
# Note: First run will download model weights from Facebook Research

# Vector Database
# ============================================================================
chromadb>=0.4.0,<0.5.0
# Note: ChromaDB has specific Python version requirements (3.10-3.12)

# Web Scraping
# ============================================================================
playwright>=1.40.0,<2.0.0
beautifulsoup4>=4.12.0,<5.0.0
lxml>=4.9.0,<5.0.0
aiohttp>=3.9.0,<4.0.0

# Web Interface
# ============================================================================
streamlit>=1.28.0,<2.0.0

# Configuration & Validation
# ============================================================================
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
pyyaml>=6.0,<7.0

# Image Processing
# ============================================================================
Pillow>=10.0.0,<11.0.0

# Utilities
# ============================================================================
python-dotenv>=1.0.0,<2.0.0
tqdm>=4.66.0,<5.0.0
colorlog>=6.8.0,<7.0.0
numpy>=1.24.0,<2.0.0

# System Monitoring
# ============================================================================
psutil>=5.9.0,<6.0.0

# Development & Testing (Optional - uncomment for development)
# ============================================================================
# pytest>=7.4.0,<8.0.0
# pytest-cov>=4.1.0,<5.0.0
# pytest-asyncio>=0.21.0,<1.0.0
# pytest-mock>=3.11.0,<4.0.0

# Jupyter Notebooks (Optional - uncomment for experimentation)
# ============================================================================
# jupyter>=1.0.0,<2.0.0
# matplotlib>=3.7.0,<4.0.0
# seaborn>=0.12.0,<1.0.0
# scikit-learn>=1.3.0,<2.0.0
# umap-learn>=0.5.0,<1.0.0

# Code Quality (Optional - uncomment for development)
# ============================================================================
# black>=23.0.0,<24.0.0
# ruff>=0.1.0,<1.0.0
# mypy>=1.7.0,<2.0.0

# ============================================================================
# Installation Notes:
# ============================================================================
# 1. Create virtual environment:
#    python -m venv venv
#    venv\Scripts\activate  (Windows PowerShell)
#
# 2. Install requirements:
#    pip install -r requirements.txt
#
# 3. Install Playwright browser:
#    playwright install chromium
#
# 4. For GPU support (Windows with CUDA 11.8):
#    pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
#
# 5. For GPU support (Windows with CUDA 12.1):
#    pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
#
# For other CUDA versions or platforms, see: https://pytorch.org/get-started/locally/
